add test cases and update user_settings.h

+++ port/posix/posix_transport_tls.h
@@ -53,7 +53,7 @@
 /** TLS configuration structure */
 typedef struct {
     char*     server_ip_string;
-    short int server_port;
+    int       server_port;
     bool      verify_peer; /* Whether to verify certificates */
 } posixTransportTlsConfig;
